+++ CHANGELOG.md
@@ -8,12 +8,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+
+### Changed
+
+## [1.0.0](https://github.com/udigeri/inout/compare/v0.9.0..v1.0.0) (2022-07-13)
+
+### Added
 - added Pay & Tokenize feature
 - added Pay with Token feature (last generated token is used for payment)
 - added Tokenize feature (just for 0,00 amount)
-- added Refund full
+- added Refund full (last generated shopping cart)
+
 ### Changed
-- some bugfix
+- bugfix about configuration of web port
+- other minor changes
 
 ## [0.9.0](https://github.com/udigeri/inout/compare/v0.8.0..v0.9.0) (2021-12-01)